pyproject.toml
[project]
name = "gaminect"
version = "0.1"
description = "Web UI for playnite data."
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE.txt"}
keywords = ["playnite", "games", "sqlite"]
authors = [
   {name = "Ivan Villareal"},
   {email = "ivan@jekiv.com"}
]
classifiers = [
   "Development Status :: 3 - Alpha",
   "Intended Audience :: Developers",
   "License :: OSI Approved :: MIT License",
   "Programming Language :: Python :: 3",
   "Programming Language :: Python :: 3.8",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3 :: Only",
   "Topic :: Games/Entertainment",
   "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
   "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "uvicorn==0.21.1",
    "fastapi==0.95.0",
    "aiohttp==3.8.4",
    "psutil==5.9.4",
    "dnspython==2.3.0",
    "asgiref==3.6.0",
    "python-json-logger==2.0.7"
]

[project.optional-dependencies]
test = [
    "pytest"
]